move merge_tensors to utils.py
